fix: rename files and use vite config in start script
