+++ src/ui/components/CustomTabs/CustomTabs.jsx
@@ -6,9 +6,9 @@ import PropTypes from 'prop-types';
 import {makeStyles} from '@material-ui/core/styles';
 import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
-import Card from '../Card/Card.js';
-import CardBody from '../Card/CardBody.js';
-import CardHeader from '../Card/CardHeader.js';
+import Card from '../Card/Card.jsx';
+import CardBody from '../Card/CardBody.jsx';
+import CardHeader from '../Card/CardHeader.jsx';
 
 import styles from '../../assets/jss/material-dashboard-react/components/customTabsStyle';
