+++ src/ui/components/Navbars/AdminNavbarLinks.jsx
@@ -13,7 +13,7 @@ import Poppers from '@material-ui/core/Popper';
 import Divider from '@material-ui/core/Divider';
 import Person from '@material-ui/icons/Person';
 import Notifications from '@material-ui/icons/Notifications';
-import Button from '../CustomButtons/Button.js';
+import Button from '../CustomButtons/Button.jsx';
 import styles from '../../assets/jss/material-dashboard-react/components/headerLinksStyle.js';
 
 const useStyles = makeStyles(styles);
