+++ src/ui/components/Navbars/Navbar.jsx
@@ -9,8 +9,8 @@ import Toolbar from '@material-ui/core/Toolbar';
 import IconButton from '@material-ui/core/IconButton';
 import Hidden from '@material-ui/core/Hidden';
 import Menu from '@material-ui/icons/Menu';
-import AdminNavbarLinks from './AdminNavbarLinks.js';
-import Button from '../CustomButtons/Button.js';
+import AdminNavbarLinks from './AdminNavbarLinks.jsx';
+import Button from '../CustomButtons/Button.jsx';
 import styles from '../../assets/jss/material-dashboard-react/components/headerStyle.js';
 
 const useStyles = makeStyles(styles);
