+++ src/ui/components/Snackbar/SnackbarContent.jsx
