+++ src/ui/views/Dashboard/Dashboard.jsx
@@ -1,8 +1,8 @@
 /* eslint-disable max-len */
 /* eslint-disable require-jsdoc */
 import React from 'react';
-import GridItem from '../../components/Grid/GridItem.js';
-import GridContainer from '../../components/Grid/GridContainer.js';
+import GridItem from '../../components/Grid/GridItem.jsx';
+import GridContainer from '../../components/Grid/GridContainer.jsx';
 import OpenPushRequestSummary from './Components/OpenPushRequestSummary';
 import PullRequestSummary from './Components/PullRequestSummary';
 import RejectedPushRequestSummary from './Components/RejectedPushRequestSummary';
