+++ src/ui/views/PushDetails/PushDetails.jsx
@@ -5,14 +5,14 @@ import {Redirect} from 'react-router-dom';
 import moment from 'moment';
 import {useHistory} from 'react-router-dom';
 import Icon from '@material-ui/core/Icon';
-import GridItem from '../../components/Grid/GridItem.js';
-import GridContainer from '../../components/Grid/GridContainer.js';
-import Card from '../../components/Card/Card.js';
-import CardIcon from '../../components/Card/CardIcon.js';
-import CardBody from '../../components/Card/CardBody.js';
-import CardHeader from '../../components/Card/CardHeader.js';
-import CardFooter from '../../components/Card/CardFooter.js';
-import Button from '../../components/CustomButtons/Button.js';
+import GridItem from '../../components/Grid/GridItem.jsx';
+import GridContainer from '../../components/Grid/GridContainer.jsx';
+import Card from '../../components/Card/Card.jsx';
+import CardIcon from '../../components/Card/CardIcon.jsx';
+import CardBody from '../../components/Card/CardBody.jsx';
+import CardHeader from '../../components/Card/CardHeader.jsx';
+import CardFooter from '../../components/Card/CardFooter.jsx';
+import Button from '../../components/CustomButtons/Button.jsx';
 import Diff from './components/Diff';
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
