+++ src/ui/views/RepoDetails/Components/AddUser.jsx
@@ -6,12 +6,12 @@ import PropTypes from 'prop-types';
 import InputLabel from '@material-ui/core/InputLabel';
 import FormControl from '@material-ui/core/FormControl';
 import FormHelperText from '@material-ui/core/FormHelperText';
-import GridItem from '../../../components/Grid/GridItem.js';
-import GridContainer from '../../../components/Grid/GridContainer.js';
-import Card from '../../../components/Card/Card.js';
-import CardBody from '../../../components/Card/CardBody.js';
+import GridItem from '../../../components/Grid/GridItem.jsx';
+import GridContainer from '../../../components/Grid/GridContainer.jsx';
+import Card from '../../../components/Card/Card.jsx';
+import CardBody from '../../../components/Card/CardBody.jsx';
 import MenuItem from '@material-ui/core/MenuItem';
-import Button from '../../../components/CustomButtons/Button.js';
+import Button from '../../../components/CustomButtons/Button.jsx';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import Select from '@material-ui/core/Select';
 import Dialog from '@material-ui/core/Dialog';
