+++ src/ui/views/RepoList/Components/TabList.jsx
@@ -1,8 +1,8 @@
 /* eslint-disable max-len */
 /* eslint-disable require-jsdoc */
 import React from 'react';
-import GridItem from '../../../components/Grid/GridItem.js';
-import GridContainer from '../../../components/Grid/GridContainer.js';
+import GridItem from '../../../components/Grid/GridItem.jsx';
+import GridContainer from '../../../components/Grid/GridContainer.jsx';
 import CustomTabs from '../../../components/CustomTabs/CustomTabs';
 import Repositories from './Repositories';
