+++ src/ui/views/UserList/Components/UserList.jsx
@@ -3,8 +3,8 @@
 import React, {useState, useEffect} from 'react';
 import {makeStyles} from '@material-ui/core/styles';
 import {useHistory} from 'react-router-dom';
-import GridItem from '../../../components/Grid/GridItem.js';
-import GridContainer from '../../../components/Grid/GridContainer.js';
+import GridItem from '../../../components/Grid/GridItem.jsx';
+import GridContainer from '../../../components/Grid/GridContainer.jsx';
 import Button from '@material-ui/core/Button';
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
