+++ vite.config.js
@@ -1,10 +1,14 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import { resolve } from 'path';
 
 export default defineConfig(() => {
   return {
     build: {
       outDir: 'build',
+    },
+    server: {
+      port: 3000,
     },
     plugins: [react()],
   };
